pyproject.toml
[tool.poetry]
name = "pyyaml_test"
version = "0.1.0"
description = ""
authors = ["Kenta Fukuhara <105337464+KAkikakuFukuhara@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
PyYAML = "^6.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
